fix: 컴포넌트 네이밍 변경
