+++ src/app/_component/modal2/todo/detail/_component/Comment/index.tsx
@@ -1,0 +1,143 @@
+'use client';
+import Image from 'next/image';
+import React, { useEffect, useState } from 'react';
+import { useSetRecoilState } from 'recoil';
+import { commentsStateAboutCardId } from '@/src/app/_recoil/CardAtom';
+import formatTime from '@/src/app/_util/formatTime';
+import { axiosInstance } from '@/src/app/_util/axiosInstance';
+import DefaultProfile from '@/src/app/_component/DefaultProfile';
+import { UserDataType } from '@/src/app/_constant/type';
+
+export interface CommentType2 {
+  id: number;
+  content: string;
+  createdAt: string;
+  updatedAt: string;
+  cardId: number;
+  author: {
+    profileImageUrl: string;
+    nickname: string;
+    id: number;
+  };
+}
+export default function Comment({ data, cardId }: { data: CommentType2; cardId: number }) {
+  // const userInfo = useRecoilValue(userInfoState);
+  // const { id:  } = userInfo;
+  const [userId, setUserId] = useState<number | null>(null);
+  const [value, setValue] = useState(data ? data.content : '');
+  const [isUpdate, setIsUpdate] = useState(false);
+  const setComments = useSetRecoilState(commentsStateAboutCardId(cardId));
+
+  const updateComments = async (e: React.MouseEvent<HTMLElement>) => {
+    e.preventDefault();
+    const res = await axiosInstance.put(`comments/${data.id}`, {
+      content: value,
+    });
+
+    setComments((oldComments: CommentType2[]) => {
+      if (oldComments) {
+        return oldComments.map((comment) => (comment?.id === data?.id ? { ...res.data } : comment));
+      }
+      return null;
+    });
+    setIsUpdate(false);
+  };
+
+  const deleteComments = async () => {
+    await axiosInstance.delete(`comments/${data.id}`);
+    setComments((oldComments: CommentType2[]) =>
+      oldComments ? oldComments.filter((comment) => comment?.id !== data?.id) : [],
+    );
+  };
+
+  const handleRenderUpdateComment = () => {
+    setIsUpdate(true);
+  };
+
+  const handleOnBlur = (e: React.FocusEvent<HTMLElement>) => {
+    e.stopPropagation();
+    setValue((e.target as HTMLInputElement).value);
+  };
+
+  const userDataObject = localStorage.getItem('taskifyUserData');
+  useEffect(() => {
+    if (userDataObject) {
+      const userData: UserDataType = JSON.parse(userDataObject);
+      const loginId = userData.userInfo.id;
+      setUserId(loginId);
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
+  if (!data) return;
+
+  return (
+    <div className='mt-4 flex gap-[0.625rem] md:mt-5'>
+      <div className='flex flex-col items-start'>
+        {data?.author?.profileImageUrl ? (
+          <div className='relative rounded-full border sm:h-[2.125rem] sm:w-[2.125rem] sm:text-[0.875rem] md:h-[2.375rem] md:w-[2.375rem]'>
+            <Image
+              src={data?.author?.profileImageUrl}
+              alt='댓글 프로필'
+              priority
+              fill
+              style={{ borderRadius: '50%' }}
+            />
+          </div>
+        ) : (
+          <DefaultProfile nickName={data?.author?.nickname} index={data?.author?.id} />
+        )}
+      </div>
+      <div className='flex w-full flex-col gap-[0.375rem]'>
+        <div className='flex gap-[0.5rem]'>
+          <span className='text-[0.875rem] font-semibold text-black'>{data?.author?.nickname}</span>
+          <span className='text-[0.75rem] text-[#9fa6b2]'>{formatTime(data && (data.createdAt as string))}</span>
+        </div>
+        {isUpdate ? (
+          <div className='my-2 flex flex-col items-center gap-[0.375rem] md:w-[37.5rem] lg:w-[40.625rem]'>
+            <input
+              id='content2'
+              type='text'
+              className='w-full border-b-2  border-black focus:outline-none'
+              placeholder='댓글을 입력해 주세요'
+              defaultValue={data.content}
+              onBlur={handleOnBlur}
+            />
+            <div className='flex w-full justify-end gap-2'>
+              <button
+                type='button'
+                className='w-14 rounded-full px-2 py-1 text-black hover:bg-gray20'
+                onClick={() => {
+                  setIsUpdate(false);
+                }}
+              >
+                취소
+              </button>
+              <button className='w-14 rounded-full bg-violet px-2 py-1   text-white' onClick={updateComments}>
+                수정
+              </button>
+            </div>
+          </div>
+        ) : (
+          <>
+            <p className='text-[0.875rem] text-black'>{data?.content}</p>
+            <div className='flex gap-[0.75rem]'>
+              {data?.author?.id === (userId as number) ? (
+                <>
+                  <span className='text-[0.75rem] text-[#9fa6b2] underline' onClick={handleRenderUpdateComment}>
+                    수정
+                  </span>
+                  <span className='text-[0.75rem] text-[#9fa6b2] underline' onClick={deleteComments}>
+                    삭제
+                  </span>
+                </>
+              ) : null}
+            </div>
+          </>
+        )}
+      </div>
+    </div>
+  );
+}
+
+// 댓글 수정 누르면 해당 댓글 부분의 내용영역만 input 필드로 바꿔야될까 음... input 위치가 좀 애매하네요
