+++ src/app/_component/modal2/todo/type.ts
@@ -14,3 +14,16 @@ export interface ToDoCardDetailProps {
   imageUrl: string;
   assignee: { profileImageUrl: string; nickname: string; id: number };
 }
+
+export interface CommentType {
+  id: number;
+  content: string;
+  createdAt: string;
+  updatedAt: string;
+  cardId: number;
+  author: {
+    profileImageUrl: string;
+    nickname: string;
+    id: number;
+  };
+}
